+++ packages/client/wallets/smart-wallet/src/api/CrossmintWalletService.ts
@@ -5,9 +5,9 @@ import { blockchainToChainId } from "@crossmint/common-sdk-base";
 
 import type { SmartWalletChain } from "../blockchain/chains";
 import { CrossmintServiceError } from "../error";
-import type { SmartWalletConfig, StoreSmartWalletParams } from "../types/api";
 import type { UserParams } from "../types/params";
 import { SmartWalletConfigSchema } from "../types/schema";
+import type { SmartWalletConfig, StoreSmartWalletParams } from "../types/service";
 import { bigintsToHex, parseBigintAPIResponse } from "../utils/api";
 import { API_VERSION } from "../utils/constants";
 import { BaseCrossmintService } from "./BaseCrossmintService";
