+++ packages/client/wallets/smart-wallet/src/blockchain/wallets/account/config.ts
@@ -3,9 +3,9 @@ import { Address } from "viem";
 import type { CrossmintWalletService } from "../../../api/CrossmintWalletService";
 import type { SmartWalletChain } from "../../../blockchain/chains";
 import { SmartWalletError } from "../../../error";
-import type { SignerData } from "../../../types/api";
 import type { SupportedEntryPointVersion, SupportedKernelVersion } from "../../../types/internal";
 import type { UserParams } from "../../../types/params";
+import type { SignerData } from "../../../types/service";
 import { EOASignerConfig, PasskeySignerConfig, SignerConfig } from "./signer";
 
 export class AccountConfigFacade {
